test: split matcher types out
