+++ apps/docs/tsconfig.json
@@ -6,7 +6,7 @@
       "@/styles/*": ["styles/*"],
       "@/components/*": ["components/*"],
       "@/data/*": ["data/*"],
-      "@modtree/test-env/jest": ["../../libs/test-env/src/jest-matchers.ts"]
+      "@modtree/test-env/jest": ["../../libs/test-env/src/matcher-types.ts"]
     },
     "allowJs": true,
     "strict": false,
