+++ jest.preset.js
@@ -3,5 +3,5 @@ const join = require('path').join
 
 module.exports = {
   ...nxPreset,
-  setupFilesAfterEnv: [join(__dirname, 'libs/test-env/src/jest-matchers.ts')],
+  setupFilesAfterEnv: [join(__dirname, 'libs/test-env/src/jest.ts')],
 }
