+++ tsconfig.base.json
@@ -26,7 +26,7 @@
       "@modtree/repo-user": ["libs/repo-user/src/index.ts"],
       "@modtree/sql": ["libs/sql/src/index.ts"],
       "@modtree/test-env": ["libs/test-env/src/index.ts"],
-      "@modtree/test-env/jest": ["libs/test-env/src/jest-matchers.ts"],
+      "@modtree/test-env/jest": ["libs/test-env/src/matcher-types.ts"],
       "@modtree/typeorm-config": ["libs/typeorm-config/src/index.ts"],
       "@modtree/types": ["libs/types/src/index.ts"],
       "@modtree/utils": ["libs/utils/src/index.ts"]
